+++ tests/conftest.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import pytest
+import torch
 from transformers.models.llama.configuration_llama import LlamaConfig
 from transformers.models.llama.modeling_llama import LlamaForCausalLM
 
@@ -16,6 +17,23 @@ from mttl.models.library.expert import Expert
 from mttl.models.library.expert_library import DatasetLibrary
 from mttl.models.modifiers.base import ModifierConfig
 from mttl.models.modifiers.lora import LoRAConfig
+
+
+@pytest.fixture
+def dummy_batch():
+    torch.manual_seed(0)
+    bs = 2
+    max_seq_len = 3
+    batch = {
+        "input_ids": torch.randint(10, 400, (bs, max_seq_len)),
+        "labels": torch.randint(10, 400, (bs, max_seq_len)),
+    }
+    seq_len = torch.randint(0, max_seq_len, (bs,))
+    attn_mask = torch.zeros(bs, max_seq_len, dtype=torch.int32)
+    attn_mask[torch.arange(bs), seq_len] = 1
+    attn_mask = 1 - attn_mask.cumsum(dim=-1)
+    batch["attention_mask"] = attn_mask
+    return batch
 
 
 @pytest.fixture(scope="session")
